added tracking of serve events
